test: add size optimization tests and add long types
- Update `tests/CMakeLists.txt` to include `float_qsort` test cases compiled with `-Os` and `-Oz` to verify DWARF generation under size optimizations.
- Update `tests/nested_structs.c` to use fixed-width integer types (`uint32_t` and `uint64_t`).

+++ tests/nested_structs.c
@@ -93,8 +93,8 @@ int main(void) {
         .stroke = {10, 20, 30, 128}
     };
 
-    int r1 = process_rectangle(&local_rect);
-    int r2 = process_complex(&global_object);
+    uint32_t r1 = process_rectangle(&local_rect);
+    uint64_t r2 = process_complex(&global_object);
 
     printf("Results: %d, %d\n", r1, r2);
     return 0;
